Moved all barcode types to own class
