+++ src/Types/TypeCode128A.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Picqer\Barcode\Types;
+
+use Picqer\Barcode\Exceptions\InvalidCharacterException;
+use Picqer\Barcode\Exceptions\InvalidLengthException;
+
+/*
+ * C128 barcodes.
+ * Very capable code, excellent density, high reliability; in very wide use world-wide
+ *
+ * @param $code (string) code to represent.
+ * @param $type (string) barcode type: A, B, C or empty for automatic switch (AUTO mode)
+ */
+
+class TypeCode128A extends TypeCode128
+{
+    protected $type = 'A';
+}
